Update money.py
Working on resolving discrepancy at end of month. Currently it calculates based on number in the month, so a payday on the 30th won't be seen as the one to pay for a bill on the first. Either move to datetime objects for the 'bills(dict)' or convert to julian days and create all 12 annual occurrances (nested inside the bills dict). Then it would be able to say 'bill X' is due from check(date early in month) and check(date late in month) rather than not catching until after the fact.